Declare the enrollment algorithm for a key
An operation on a key may indifferently use the algorithm set with
psa_set_key_algorithm() or with this function.

Warning:
Setting an enrollment algorithm is not recommended, because
using the same key with different algorithms can allow some
attacks based on arithmetic relations between different
computations made with the same key, or can escalate harmless
side channels into exploitable ones. Use this function only
if it is necessary to support a protocol for which it has been
verified that the usage of the key with multiple algorithms
is safe.

Signed-off-by: Devaraj Ranganna <devaraj.ranganna@arm.com>